Refactor: Add custom hook for getUserData
