+++ jsconfig.json
@@ -6,6 +6,7 @@
       "assets/*": ["src/assets/*"],
       "components/*": ["src/components/*"],
       "contexts/*": ["src/contexts/*"],
+      "hooks/*": ["src/hooks/*"],
       "pages/*": ["src/pages/*"],
       "utils/*": ["src/utils/*"]
     }
