+++ src/components/Footer.jsx
@@ -1,4 +1,4 @@
-import { Link, useLocation } from "react-router-dom";
+import { Link } from "react-router-dom";
 import styled from "styled-components";
 import FacebookIcon from "assets/icon-facebook.png";
 import TwitterIcon from "assets/icon-twitter.png";
@@ -58,11 +58,6 @@ const IconBox = styled.div`
 `;
 
 function Footer() {
-  const location = useLocation();
-  if (["/signin", "/signup"].includes(location.pathname)) {
-    return null;
-  }
-
   return (
     <FooterWrapper>
       <FooterContainer>
