+++ src/components/Nav.jsx
@@ -1,10 +1,8 @@
-import { useEffect, useState } from "react";
-import { Link, useLocation } from "react-router-dom";
+import { Link } from "react-router-dom";
 import styled from "styled-components";
 import LinkButton from "components/LinkButton";
 import { useUserId } from "contexts/UserIdContext";
-import { getUsers } from "utils/api";
-import DefaultProfileSource from "assets/default-profile.png";
+import useUserData from "hooks/useUserData";
 
 const NavWrapper = styled.nav`
   position: sticky;
@@ -84,28 +82,7 @@ const ProfileEmail = styled.p`
 `;
 
 function Nav() {
-  const location = useLocation();
-  if (["/signin", "/signup"].includes(location.pathname)) {
-    return;
-  }
-
-  const [userEmail, setUserEmail] = useState("");
-  const [userImage, setUserImage] = useState(DefaultProfileSource);
-
-  // userId에 맞는 데이터를 가져와야 하는데
-  // api에 데이터가 하나밖에 없고 배열의 형태가 아니라서
-  // 일단 조건 없이 가져오는 것으로 임시 처리했습니다.
-  const getUserData = async () => {
-    const { data } = await getUsers();
-    if (!data) return;
-    const { email, profileImageSource } = data;
-    setUserEmail(email || "");
-    setUserImage(profileImageSource || DefaultProfileSource);
-  };
-
-  useEffect(() => {
-    getUserData();
-  }, []);
+  const { userEmail, userImage } = useUserData();
 
   const userId = useUserId();
   const isAuth = userId > 0 ? true : false;
