+++ vite.config.js
@@ -11,6 +11,7 @@ export default defineConfig({
       assets: path.resolve(__dirname, "src/assets"),
       components: path.resolve(__dirname, "src/components"),
       contexts: path.resolve(__dirname, "src/contexts"),
+      hooks: path.resolve(__dirname, "src/hooks"),
       pages: path.resolve(__dirname, "src/pages"),
       utils: path.resolve(__dirname, "src/utils"),
     },
